% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNlWasteDistrib.R
\name{calcNlWasteDistrib}
\alias{calcNlWasteDistrib}
\title{calcNlWasteDistrib}
\usage{
calcNlWasteDistrib()
}
\value{
Magpie object of waste types to waste distribution,  share
}
\description{
non-linear optimization distributes waste by composition type to disposal type.
returns magpie object, share of total disposal
}
\examples{

\dontrun{ a <- calcOutput(type="NlWasteDistrib")
}
}
\author{
David Chen
}
