% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcOceanPlastic.R
\name{calcOceanPlastic}
\alias{calcOceanPlastic}
\title{calcOceanPlastic}
\usage{
calcOceanPlastic(filtration = 0.4)
}
\arguments{
\item{filtration}{amount of plastic NOT captured by filtration, janbeck uses 0.15, 0.25, 0.4, but cites a source at .6}
}
\value{
million kg of plastic waste
}
\description{
Calculates amount of ocean plastic based on calibrated waste projections, 
for unmanaged waste: all dumps, landfills and dumps from develolping countries, and 2% total as litter
Coastal pop from Janbeck et al. 2016 buffer, as constant percentage future pop
}
\author{
David Chen
}
