% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWasteGen.R
\name{calcWasteGen}
\alias{calcWasteGen}
\title{calcWasteGen}
\usage{
calcWasteGen(pc = TRUE, form = "LogLog")
}
\arguments{
\item{pc}{per capita (kg/capita) or total (Mt)}

\item{form}{Functional form of predicted waste generation}
}
\value{
magpie object of total waste generation
}
\description{
Calculates waste generation based on WhataWaste2.0 data, based on gdp regressions
and calibrated to real data multiplicatively
}
\author{
David Chen
}
