% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWasteTrt.R
\name{calcWasteTrt}
\alias{calcWasteTrt}
\title{calcWasteDirTrt}
\usage{
calcWasteTrt(weight = "pop", SSP = "SSP2")
}
\arguments{
\item{weight}{population weights or "none"}

\item{SSP}{SSP scenario}
}
\value{
magpie object of waste treatment by type share
}
\description{
Calculates shares of waste treatments by type based on Dirichlet regression on gdp using WhataWaste2.0 data
note that each type is independent - treatments for each type all sum to 1
}
\author{
David Chen
}
