% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWasteType.R
\name{calcWasteType}
\alias{calcWasteType}
\title{calcWasteType}
\usage{
calcWasteType(weight = "pop", SSP = "SSP2")
}
\arguments{
\item{weight}{population weights or other weights or NULL}

\item{SSP}{SSP scenario}
}
\value{
magpie object of waste shares
}
\description{
Calculates shares of waste types based on Dirichlet regression on gdp using WhataWaste2.0 data
}
\author{
David Chen
}
