% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCoastalPop.R
\name{convertCoastalPop}
\alias{convertCoastalPop}
\title{convertCoastalPop}
\usage{
convertCoastalPop()
}
\value{
pop in millions
}
\description{
Fills and completes Janbeck 2011 coastal population data for all years based on constant percentage of coastal pop for future years
}
\seealso{
\code{\link{readSource}}
}
\author{
David Chen
}
