% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrwaste-package.R
\docType{package}
\name{mrwaste-package}
\alias{mrwaste}
\alias{mrwaste-package}
\title{mrwaste: Analysis and Projection of Municipal Solid Waste}
\description{
Reads in waste data from What a Waste 2.0 and uses brms package to create future regressions based on GDP.
}
\author{
\strong{Maintainer}: David Meng-Chuen Chen \email{david.chen@pik-potsdam.de}

}
\keyword{internal}
