% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCoastalPop.R
\name{readCoastalPop}
\alias{readCoastalPop}
\title{Read Janbeck 2011 Coastal Population data, for Ocean Plastic input}
\usage{
readCoastalPop()
}
\value{
magpie object of coastal population by country, 50km buffer around coastline,
keeping percentage same in future
}
\description{
Read-in a xlsx file as magclass object
kg/cap
}
\examples{

\dontrun{ a <- readSource(type="CoastalPop")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
David Chen
}
